cleaning up the authentication state compose screens
